+++ CHANGES.md
@@ -19,6 +19,7 @@
 - Added hover tooltips to inline links
 - Links can now be copied
 - Add ability to link local files directly from notes (#24)
+  - You can opt into double clicking in settings if you prefer
 
 ## ✨ Improvements and fixes
 - Fix code blocks breaking when all content is deleted with ⌘+⌫ ⌘+Del, Ctrl+⌫ or Ctrl+Del